requirements.txt
Flask==2.3.4
groq
python-dotenv
